Tab Refactor: Account Landing, OBJ Landing, FW Details (#6375)
* updating account landing, FW detail and OBJ landing to new linked tab pattern

* fixing index for key issue
