+++ packages/manager/src/features/Account/EnableManaged.tsx
@@ -18,7 +18,6 @@ import { pluralize } from 'src/utilities/pluralize';
 interface Props {
   isManaged: boolean;
   update: (data: Partial<AccountSettings>) => void;
-  push: (url: string) => void;
 }
 
 interface StateProps {
@@ -68,7 +67,7 @@ export const ManagedContent: React.FC<ContentProps> = props => {
 };
 
 export const EnableManaged: React.FC<CombinedProps> = props => {
-  const { isManaged, linodeCount, push, update } = props;
+  const { isManaged, linodeCount, update } = props;
   const [isOpen, setOpen] = React.useState<boolean>(false);
   const [error, setError] = React.useState<string | undefined>();
   const [isLoading, setLoading] = React.useState<boolean>(false);
@@ -90,7 +89,7 @@ export const EnableManaged: React.FC<CombinedProps> = props => {
       .then(() => {
         handleClose();
         update({ managed: true });
-        push('/managed');
+        //push('/managed');
       })
       .catch(handleError);
   };
